Safely handle recursive types
