+++ infer/resource.go
@@ -16,9 +16,7 @@ package infer
 
 import (
 	"fmt"
-	"os"
 	"reflect"
-	"strconv"
 
 	"github.com/hashicorp/go-multierror"
 	pschema "github.com/pulumi/pulumi/pkg/v3/codegen/schema"
@@ -1013,280 +1011,4 @@ func insertSecrets(props resource.PropertyMap, secrets []resource.PropertyPath) 
 		s.Set(m, resource.MakeSecret(v))
 	}
 	return m.ObjectValue()
-}
-
-// applyDefaults recursively applies the default values provided by [introspect.Annotator].
-func applyDefaults[T any](value *T) error {
-	v := reflect.ValueOf(value).Elem()
-	contract.Assertf(v.CanSet(), "Cannot accept an un-editable pointer")
-
-	// Defaults can only sit on objects, but objects can be under any level of
-	// indirection.
-	//
-	// We thus need to walk the entire structure to find defaults to apply.
-	var walk func(v reflect.Value) (didSet bool, _ error)
-
-	// apply may only be called on structs.
-	apply := func(v reflect.Value) (bool, error) {
-		t := v.Type()
-		a := getAnnotated(t)
-
-		// If v is a nil valued struct with defaults, we should hydrate it and
-		// apply the default. If not, we leave it nil. This must be applied
-		// recursively.
-		fields := map[string]reflect.Value{}
-		for _, field := range reflect.VisibleFields(v.Type()) {
-			tag, err := introspect.ParseTag(field)
-			if err != nil {
-				return false, err
-			}
-			if tag.Internal {
-				continue
-			}
-			fields[tag.Name] = v.FieldByIndex(field.Index)
-		}
-
-		var didSet bool
-	defaultEnvs:
-		for k, envVars := range a.DefaultEnvs {
-			value, ok := fields[k]
-			if ok && !value.IsZero() {
-				continue
-			}
-			for _, env := range envVars {
-				envValue := os.Getenv(env)
-				if envValue == "" {
-					continue
-				}
-				err := setDefaultValueFromEnv(value, envValue)
-				if err != nil {
-					return false, err
-				}
-				didSet = true
-				continue defaultEnvs
-			}
-		}
-
-		for k, inMemoryDefault := range a.Defaults {
-			value, ok := fields[k]
-			if ok && !value.IsZero() {
-				continue
-			}
-			err := setDefaultFromMemory(value, inMemoryDefault)
-			if err != nil {
-				return false, err
-			}
-			didSet = true
-		}
-
-		for _, v := range fields {
-			didSetRec, err := walk(v)
-			if err != nil {
-				return false, err
-			}
-			if didSetRec {
-				didSet = true
-			}
-		}
-		return didSet, nil
-	}
-
-	// walk is responsible for calling applyDefaults on all structs in reachable from value.
-	walk = func(v reflect.Value) (didSet bool, _ error) {
-		tDeferenced := v.Type()
-		for tDeferenced.Kind() == reflect.Pointer {
-			tDeferenced = tDeferenced.Elem()
-		}
-		switch tDeferenced.Kind() {
-		// We apply defaults to each value in slice.
-		case reflect.Slice:
-			// Dereference to the underlying slice
-			for v.Kind() == reflect.Pointer && !v.IsNil() {
-				v = v.Elem()
-			}
-			// Either we have a type *[]T, **[]T, etc. and a pointer is nil,
-			// or we have reached the slice and the slice itself is nil. Both
-			// cases prevent us finding any more structs, so we are done.
-			if v.IsNil() {
-				return didSet, nil
-			}
-			for i := 0; i < v.Len(); i++ {
-				didSetRec, err := walk(v.Index(i))
-				if err != nil {
-					return false, err
-				}
-				if didSetRec {
-					didSet = true
-				}
-			}
-			return didSet, nil
-
-		case reflect.Array:
-			// Dereference to the underlying slice
-			for v.Kind() == reflect.Pointer {
-				if v.IsNil() {
-					return false, nil
-				}
-				v = v.Elem()
-			}
-
-			// Arrays cannot be nil, so we don't (and can't) perform a nil
-			// check here.
-			for i := 0; i < v.Len(); i++ {
-				didSetRec, err := walk(v.Index(i))
-				if err != nil {
-					return false, err
-				}
-				if didSetRec {
-					didSet = true
-				}
-			}
-			return didSet, nil
-
-		case reflect.Map:
-			// Dereference to the underlying map
-			for v.Kind() == reflect.Pointer && !v.IsNil() {
-				v = v.Elem()
-			}
-			// Either we have a type *map[K]V, **map[K]V, etc. and a pointer
-			// is nil, or we have reached the map and the map itself is
-			// nil. Both cases prevent us finding any more structs, so we are
-			// done.
-			if v.IsNil() {
-				return false, nil
-			}
-			iter := v.MapRange()
-			for iter.Next() {
-				value := reflect.New(iter.Value().Type()).Elem()
-				value.Set(iter.Value())
-				s := hydratedValue(value)
-				didSetRec, err := walk(s)
-				if err != nil {
-					return false, err
-				}
-				if didSetRec {
-					v.SetMapIndex(iter.Key(), s)
-					didSet = true
-				}
-			}
-			return didSet, nil
-		case reflect.Struct:
-			// We work on a copy because we don't want to hydrate an actual
-			// field if we don't every apply a default to it.
-			copy := reflect.New(v.Type()).Elem()
-			copy.Set(v)
-			s := hydratedValue(copy)
-			didSet, err := apply(derefNonNil(s))
-			if err != nil {
-				return false, err
-			}
-			if didSet {
-				v.Set(s)
-			}
-			return didSet, nil
-
-		// This is a primitive type. That means that:
-		//
-		// 1. It is not a struct.
-		// 2. It cannot contain a struct.
-		//
-		// That means there are not any defaults to apply, so we're done.
-		default:
-			return false, nil
-		}
-	}
-
-	_, err := walk(v)
-	return err
-}
-
-// setDefaultFromMemory applies an in-memory default value to a field that can accept one.
-//
-// field must be CanSet and value must either be a primitive, or point to one.
-func setDefaultFromMemory(field reflect.Value, value any) error {
-	if value == nil {
-		return nil
-	}
-	// We will set field to a primitive value, we can freely provide hydration.
-	field.Set(hydratedValue(field))
-	field = derefNonNil(field)
-
-	v := reflect.ValueOf(value)
-	if v.CanConvert(field.Type()) {
-		field.Set(v.Convert(field.Type()))
-		return nil
-	}
-	return fmt.Errorf("cannot set field of type '%s' to default value %q (%[2]T)",
-		field.Type(), value)
-}
-
-// setDefaultValueFromEnv applies a default value source from an environmental variable to
-// a field.
-//
-// field must be CanSet and value must either be a primitive, or point to one.
-func setDefaultValueFromEnv(field reflect.Value, value string) error {
-	typ := field.Type()
-	for typ.Kind() == reflect.Pointer {
-		typ = typ.Elem()
-	}
-
-	switch typ.Kind() {
-	case reflect.String:
-		return setDefaultFromMemory(field, value)
-	case reflect.Float64:
-		f, err := strconv.ParseFloat(value, 64)
-		if err != nil {
-			return err
-		}
-		return setDefaultFromMemory(field, f)
-	case reflect.Int:
-		i, err := strconv.ParseInt(value, 0, 64)
-		if err != nil {
-			return err
-		}
-		return setDefaultFromMemory(field, i)
-	case reflect.Bool:
-		b, err := strconv.ParseBool(value)
-		if err != nil {
-			return err
-		}
-		return setDefaultFromMemory(field, b)
-	}
-	return fmt.Errorf("unable to apply default value %q to field of type %s",
-		value, typ)
-}
-
-// hydratedValue takes a (possibly ptr) and returns a fully hydrated value of the same
-// type. For example:
-//
-//	T{}        -> T{}
-//	T{N: 1}    -> T{N: 1}
-//	&T{N: 2}   -> &T{N: 2}
-//	(*T)(nil)  -> &T{}
-//	(**T)(nil) -> &&T{}
-func hydratedValue(value reflect.Value) reflect.Value {
-	// root := *&T where T = typeof(value)
-	//
-	// This creates an addressable value of the same type as value
-	root := value
-	for value.Kind() == reflect.Pointer {
-		// We have *T, so we need to construct a T and set the value to it.
-		if value.IsNil() {
-			// This is the reflect equivalent of:
-			//
-			//	var v typeof(value)
-			//	*elem := value
-			v := reflect.New(value.Type().Elem())
-			value.Set(v)
-		}
-		value = value.Elem()
-	}
-	return root
-}
-
-func derefNonNil(value reflect.Value) reflect.Value {
-	for value.Kind() == reflect.Pointer {
-		value = value.Elem()
-	}
-	return value
 }
